refactor code, add acumulation gradient, support efficentdetd0-d7
