--- data/coco_labels.txt
@@ -1,80 +1,0 @@
-1,1,person
-2,2,bicycle
-3,3,car
-4,4,motorcycle
-5,5,airplane
-6,6,bus
-7,7,train
-8,8,truck
-9,9,boat
-10,10,traffic light
-11,11,fire hydrant
-13,12,stop sign
-14,13,parking meter
-15,14,bench
-16,15,bird
-17,16,cat
-18,17,dog
-19,18,horse
-20,19,sheep
-21,20,cow
-22,21,elephant
-23,22,bear
-24,23,zebra
-25,24,giraffe
-27,25,backpack
-28,26,umbrella
-31,27,handbag
-32,28,tie
-33,29,suitcase
-34,30,frisbee
-35,31,skis
-36,32,snowboard
-37,33,sports ball
-38,34,kite
-39,35,baseball bat
-40,36,baseball glove
-41,37,skateboard
-42,38,surfboard
-43,39,tennis racket
-44,40,bottle
-46,41,wine glass
-47,42,cup
-48,43,fork
-49,44,knife
-50,45,spoon
-51,46,bowl
-52,47,banana
-53,48,apple
-54,49,sandwich
-55,50,orange
-56,51,broccoli
-57,52,carrot
-58,53,hot dog
-59,54,pizza
-60,55,donut
-61,56,cake
-62,57,chair
-63,58,couch
-64,59,potted plant
-65,60,bed
-67,61,dining table
-70,62,toilet
-72,63,tv
-73,64,laptop
-74,65,mouse
-75,66,remote
-76,67,keyboard
-77,68,cell phone
-78,69,microwave
-79,70,oven
-80,71,toaster
-81,72,sink
-82,73,refrigerator
-84,74,book
-85,75,clock
-86,76,vase
-87,77,scissors
-88,78,teddy bear
-89,79,hair drier
-90,80,toothbrush
